requirements.txt
matplotlib
numpy
tk
sympy
pylatex
pillow
